dashboard/requirements.txt
streamlit
pandas
altair
python-dotenv
redshift_connector
pytest
pylint